OCR_test/requirements.txt
anyio==4.6.0
astor==0.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
easyocr==1.7.2
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
fsspec==2024.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.35.1
imgaug==0.4.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==3.20.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pyparsing==3.1.4
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.3
termcolor==2.5.0
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
